Fix review comment

+++ velox/functions/sparksql/tests/DateTimeFunctionsTest.cpp
@@ -970,7 +970,7 @@ TEST_F(DateTimeFunctionsTest, makeYMInterval) {
 
 TEST_F(DateTimeFunctionsTest, yearOfWeek) {
   const auto yearOfWeek = [&](std::optional<int32_t> date) {
-    return evaluateOnce<int64_t, int32_t>("year_of_week(c0)", {date}, {DATE()});
+    return evaluateOnce<int32_t, int32_t>("year_of_week(c0)", {date}, {DATE()});
   };
   EXPECT_EQ(std::nullopt, yearOfWeek(std::nullopt));
   EXPECT_EQ(1970, yearOfWeek(0));
